SQL - Fix bug vue matérialisée

+++ occtax/install/sql/materialized_views.pgsql.sql
@@ -755,7 +755,7 @@ UNION
     'Nombre de producteurs ayant transmis des jeux de données' AS libelle,
     count(DISTINCT r.id_organisme) AS valeur
    FROM ( SELECT jdd.jdd_id,
-            (jsonb_array_elements(jdd.ayants_droit) ->> 'id_organisme'::text)::integer AS id_organisme,
+            jsonb_array_elements(jdd.ayants_droit) ->> 'id_organisme'::text AS id_organisme,
             jsonb_array_elements(jdd.ayants_droit) ->> 'role'::text AS role
            FROM occtax.jdd) r
 UNION
